feat(config-optimizer): extract as plugin, rename extra, move CLI from core
- Move config_optimizer to nat/plugins/config_optimizer/ (plugin layout)
- Move optimize CLI from core to nvidia-nat-config-optimizer plugin
- Rename nvidia-nat[optimizer] extra to nvidia-nat[config-optimizer]
- Remove config optimizer from default nvidia-nat install
- Update imports, docs, and examples for new paths
