+++ packages/nvidia_nat_config_optimizer/src/nat/plugins/config_optimizer/optimizer_runtime.py
@@ -19,8 +19,8 @@ import logging
 from pydantic import BaseModel
 
 from nat.cli.type_registry import GlobalTypeRegistry
-from nat.config_optimizer import register  # noqa: F401 - trigger optimizer registration
-from nat.config_optimizer.optimizable_utils import walk_optimizables
+from nat.plugins.config_optimizer import register  # noqa: F401 - trigger optimizer registration
+from nat.plugins.config_optimizer.optimizable_utils import walk_optimizables
 from nat.data_models.optimizer import OptimizerRunConfig
 from nat.experimental.decorators.experimental_warning_decorator import experimental
 from nat.runtime.loader import load_config
