+++ packages/nvidia_nat_config_optimizer/src/nat/plugins/config_optimizer/register.py
@@ -18,9 +18,9 @@ import asyncio
 from collections.abc import AsyncIterator
 
 from nat.cli.register_workflow import register_optimizer
-from nat.config_optimizer.parameters.base import BaseParameterOptimizer
-from nat.config_optimizer.parameters.optimizer import optimize_parameters
-from nat.config_optimizer.prompts.ga_prompt_optimizer import GAPromptOptimizer
+from nat.plugins.config_optimizer.parameters.base import BaseParameterOptimizer
+from nat.plugins.config_optimizer.parameters.optimizer import optimize_parameters
+from nat.plugins.config_optimizer.prompts.ga_prompt_optimizer import GAPromptOptimizer
 from nat.data_models.config import Config
 from nat.data_models.optimizable import SearchSpace
 from nat.data_models.optimizer import GAPromptOptimizationConfig
