+++ packages/nvidia_nat_config_optimizer/tests/test_parameter_selection_extra.py
@@ -16,7 +16,7 @@
 import optuna
 from optuna.study import StudyDirection
 
-from nat.config_optimizer.parameters.selection import pick_trial
+from nat.plugins.config_optimizer.parameters.selection import pick_trial
 
 
 def _make_study_with_trials(values_list):  # noqa: ANN001
