+++ packages/nvidia_nat_config_optimizer/tests/test_pareto_visualizer_extra.py
@@ -18,7 +18,7 @@ from pathlib import Path
 import optuna
 import pandas as pd
 
-from nat.config_optimizer.parameters.pareto_visualizer import create_pareto_visualization
+from nat.plugins.config_optimizer.parameters.pareto_visualizer import create_pareto_visualization
 
 
 def _make_two_obj_study():
