+++ packages/nvidia_nat_core/pyproject.toml
@@ -140,7 +140,6 @@ configure = "nat.cli.commands.configure.configure:configure_command"
 finetune = "nat.cli.commands.finetune:finetune_command"
 info = "nat.cli.commands.info.info:info_command"
 object-store = "nat.cli.commands.object_store.object_store:object_store_command"
-optimize = "nat.cli.commands.optimize:optimizer_command"
 registry = "nat.cli.commands.registry.registry:registry_command"
 start = "nat.cli.commands.start:start_command"
 uninstall = "nat.cli.commands.uninstall:uninstall_command"
