+++ packages/nvidia_nat_core/tests/nat/cli/test_plugin_loader.py
@@ -173,7 +173,6 @@ class TestPluginLoaderIntegration:
         "finetune",
         "info",
         "object-store",
-        "optimize",
         "red-team",
         "registry",
         "start",
