+++ pyproject.toml
@@ -45,7 +45,6 @@ source = "https://github.com/NVIDIA/NeMo-Agent-Toolkit"
 [tool.setuptools_dynamic_dependencies]
 dependencies = [
   "nvidia-nat-core == {version}",
-  "nvidia-nat-config-optimizer == {version}",
 ]
 
 [tool.setuptools_dynamic_dependencies.optional-dependencies]
@@ -67,7 +66,7 @@ mcp = ["nvidia-nat-mcp == {version}"]
 mem0ai = ["nvidia-nat-mem0ai == {version}"]
 nemo-customizer = ["nvidia-nat-nemo-customizer == {version}"]
 openpipe-art = ["nvidia-nat-openpipe-art == {version}"]
-optimizer = ["nvidia-nat-config-optimizer == {version}"]
+config-optimizer = ["nvidia-nat-config-optimizer == {version}"]
 opentelemetry = ["nvidia-nat-opentelemetry == {version}"]
 phoenix = ["nvidia-nat-phoenix == {version}"]
 profiler = ["nvidia-nat-profiler == {version}"]
